Fix iOS Sample App build
When following the instructions for running the sample app (iOS), it failed. I must have added the lifecycles module incorrectly, in a way where the build only worked on my machine in a particular path. Removing and readding the module changed some build artifacts and now everything seems to work. Specifically it fixed the path of some pbxfilereference.